Move keyboard teleop under devtools
