+++ ai_sapiens_sim2real/test/test_keyboard_teleop.cpp
@@ -18,7 +18,7 @@
 
 #include <yaml-cpp/yaml.h>  // NOLINT(build/include_order)
 
-#include "ai_sapiens_sim2real/keyboard_teleop/keyboard_teleop.hpp"
+#include "ai_sapiens_sim2real/teleop_devtools/keyboard/keyboard_teleop.hpp"
 
 namespace ai_sapiens_sim2real
 {
